Moves duplicate EventManager code into the core
Refs #219
