+++ WatchDogCore/WatchDog/src/nl/tudelft/watchdog/core/logic/document/EditorWrapperBase.java
@@ -4,4 +4,7 @@ package nl.tudelft.watchdog.core.logic.document;
  * Wrapper interface for IDE's editor. Each Plugin should implement it and adapt it to its API.
  */
 public interface EditorWrapperBase {
+	
+	/** Getting the current editor. */
+	Object getEditor();
 }
