+++ WatchDogCore/WatchDog/src/nl/tudelft/watchdog/core/logic/ui/EventManagerBase.java
@@ -2,12 +2,267 @@ package nl.tudelft.watchdog.core.logic.ui;
 
 import java.util.Date;
 
+import nl.tudelft.watchdog.core.logic.document.Document;
+import nl.tudelft.watchdog.core.logic.document.EditorWrapperBase;
+import nl.tudelft.watchdog.core.logic.interval.IDEIntervalManagerBase;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.EditorIntervalBase;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.IDEActiveInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.IDEOpenInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.IntervalBase;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.IntervalType;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.PerspectiveInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.PerspectiveInterval.Perspective;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.ReadingInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.TypingInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.UserActiveInterval;
+import nl.tudelft.watchdog.core.logic.interval.intervaltypes.WatchDogViewInterval;
+import nl.tudelft.watchdog.core.logic.ui.events.EditorEvent;
 import nl.tudelft.watchdog.core.logic.ui.events.WatchDogEvent;
+import nl.tudelft.watchdog.core.logic.ui.events.WatchDogEvent.EventType;
 
-public interface EventManagerBase {
+/**
+ * Manager for {@link EditorEvent}s. Links such events to actions in the
+ * IntervalManager, i.e. manages the creation and deletion of intervals based on
+ * the incoming events. This class therefore contains the logic of when and how
+ * new intervals are created, and how WatchDog reacts to incoming events
+ * generated by its listeners.
+ */
+public abstract class EventManagerBase {
 	
-	void update(WatchDogEvent event);
+	/** The {@link IntervalManager} this observer is working with. */
+	protected final IDEIntervalManagerBase intervalManager;
+
+	private final UserInactivityNotifier userInactivityNotifier;
+
+	private final InactivityNotifier typingInactivityNotifier;
+
+	private final InactivityNotifier readingInactivityNotifier;
 	
-	void update(WatchDogEvent event, Date forcedDate);
+	/** Constructor. */
+	public EventManagerBase(final IDEIntervalManagerBase intervalManager,
+			int userActivityTimeout) {
+		this.intervalManager = intervalManager;
+
+		// the user inactivity timer additionally calls a new user event
+		userInactivityNotifier = new UserInactivityNotifier(this,
+				userActivityTimeout, EventType.USER_INACTIVITY);
+		typingInactivityNotifier = new InactivityNotifier(this,
+				userActivityTimeout, EventType.TYPING_INACTIVITY);
+		readingInactivityNotifier = new InactivityNotifier(this,
+				userActivityTimeout, EventType.READING_INACTIVITY);
+	}
+
+	/**
+	 * Simple proxy for {@link #update(WatchDogEvent, Date)}, calling it with
+	 * the forcedDate set to "now".
+	 */
+	public void update(WatchDogEvent event) {
+		update(event, new Date());
+	}
+
+	/**
+	 * Introduces the supplied editorEvent, and sets the forcedDate to the
+	 * interval that is being created or closed.
+	 */
+	public void update(WatchDogEvent event, Date forcedDate) {
+		IntervalBase interval;
+		switch (event.getType()) {
+		case START_IDE:
+			intervalManager.addInterval(new IDEOpenInterval(forcedDate));
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+
+		case END_IDE:
+			userInactivityNotifier.cancelTimer(forcedDate);
+			break;
+
+		case ACTIVE_WINDOW:
+			interval = intervalManager.getInterval(IDEActiveInterval.class);
+			if (isClosed(interval)) {
+				intervalManager.addInterval(new IDEActiveInterval(forcedDate));
+			}
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+
+		case INACTIVE_WINDOW:
+			interval = intervalManager.getInterval(IDEActiveInterval.class);
+			intervalManager.closeInterval(interval, forcedDate);
+			break;
+			
+		case START_PERSPECTIVE:
+			Perspective perspecitve = (Perspective) event.getSource();
+			createNewPerspectiveInterval(perspecitve, forcedDate);
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+			
+		case JUNIT:
+			addJUnitInterval(event);
+			break;
+			
+		case USER_ACTIVITY:
+			interval = intervalManager.getInterval(UserActiveInterval.class);
+			if (isClosed(interval)) {
+				intervalManager.addInterval(new UserActiveInterval(forcedDate));
+			}
+			userInactivityNotifier.trigger();
+			break;
+
+		case START_EDIT:
+			EditorIntervalBase editorInterval = intervalManager
+					.getEditorInterval();
+			Object editor = event.getSource();
+
+			readingInactivityNotifier.cancelTimer(forcedDate);
+			if (intervalIsOfType(editorInterval, IntervalType.TYPING)
+					&& !isDifferentEditor(editorInterval, editor)) {
+				return;
+			}
+
+			intervalManager.closeInterval(editorInterval, forcedDate);
+
+			TypingInterval typingInterval = new TypingInterval(
+					createEditorWrapper(editor), forcedDate);
+			Document document = null;
+			if (editorInterval != null
+					&& !isDifferentEditor(editorInterval, editor)) {
+				document = editorInterval.getDocument();
+			} else {
+				document = createDocument(editor);
+			}
+			typingInterval.setDocument(document);
+			intervalManager.addInterval(typingInterval);
+
+			typingInactivityNotifier.trigger();
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+			
+		case SUBSEQUENT_EDIT:
+			editorInterval = intervalManager.getEditorInterval();
+			editor = event.getSource();
+
+			if (isClosed(editorInterval)
+					|| !intervalIsOfType(editorInterval, IntervalType.TYPING)
+					|| isDifferentEditor(editorInterval, editor)) {
+				update(new WatchDogEvent(event.getSource(),
+						EventType.START_EDIT));
+				break;
+			}
+
+			typingInactivityNotifier.trigger();
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+			
+		case PAINT:
+		case CARET_MOVED:
+		case ACTIVE_FOCUS:
+			editorInterval = intervalManager.getEditorInterval();
+			editor = event.getSource();
+			if (needToCreateNewReadingInterval(editorInterval, editor)) {
+				if (!isClosed(editorInterval)) {
+					intervalManager.closeInterval(editorInterval, forcedDate);
+				}
+
+				ReadingInterval readingInterval = new ReadingInterval(
+						createEditorWrapper(editor), forcedDate);
+				readingInterval
+						.setDocument(createDocument(editor));
+				intervalManager.addInterval(readingInterval);
+			}
+
+			readingInactivityNotifier.trigger();
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+			
+		case INACTIVE_FOCUS:
+			editorInterval = intervalManager.getEditorInterval();
+			intervalManager.closeInterval(editorInterval, forcedDate);
+			readingInactivityNotifier.cancelTimer(forcedDate);
+			typingInactivityNotifier.cancelTimer(forcedDate);
+			break;
+
+		case USER_INACTIVITY:
+			interval = intervalManager.getInterval(UserActiveInterval.class);
+			intervalManager.closeInterval(interval, forcedDate);
+			typingInactivityNotifier.cancelTimer(forcedDate);
+			readingInactivityNotifier.cancelTimer(forcedDate);
+			break;
+
+		case TYPING_INACTIVITY:
+			editorInterval = intervalManager.getEditorInterval();
+			if (intervalIsOfType(editorInterval, IntervalType.TYPING)) {
+				intervalManager.closeInterval(editorInterval, forcedDate);
+			}
+			break;
+
+		case READING_INACTIVITY:
+			editorInterval = intervalManager.getEditorInterval();
+			if (intervalIsOfType(editorInterval, IntervalType.READING)) {
+				intervalManager.closeInterval(editorInterval, forcedDate);
+			}
+			break;
+
+		case START_WATCHDOGVIEW:
+			interval = intervalManager.getInterval(WatchDogViewInterval.class);
+			if (!intervalIsOfType(interval, IntervalType.WATCHDOGVIEW)) {
+				intervalManager
+						.addInterval(new WatchDogViewInterval(forcedDate));
+			}
+			userInactivityNotifier.trigger(forcedDate);
+			break;
+
+		case END_WATCHDOGVIEW:
+			interval = intervalManager.getInterval(WatchDogViewInterval.class);
+			if (intervalIsOfType(interval, IntervalType.WATCHDOGVIEW)) {
+				intervalManager.closeInterval(interval, forcedDate);
+			}
+			break;
+
+		default:
+			break;
+		}
+	}
+	
+	/** Create and add a JUnitInterval based on the event. */
+	protected abstract void addJUnitInterval(WatchDogEvent event);
+	
+	/** Creates an EditorWrapper for the current editor. */
+	protected abstract EditorWrapperBase createEditorWrapper(Object editor);
+
+	/** Creates a {@link Document} based on the editor contents. */
+	protected abstract Document createDocument(Object editor);
+	
+	private boolean needToCreateNewReadingInterval(
+			EditorIntervalBase editorInterval, Object editor) {
+		return isClosed(editorInterval)
+				|| isDifferentEditor(editorInterval, editor);
+	}
+
+	private boolean isDifferentEditor(EditorIntervalBase editorInterval,
+			Object editor) {
+		return editorInterval.getEditorWrapper().getEditor() != editor;
+	}
+	
+	private boolean intervalIsOfType(IntervalBase interval, IntervalType type) {
+		return interval != null && interval.getType() == type;
+	}
+
+	private boolean isClosed(IntervalBase interval) {
+		return interval == null || interval.isClosed();
+	}
+	
+	/** Creates a new perspective Interval of the given type. */
+	private void createNewPerspectiveInterval(
+			PerspectiveInterval.Perspective perspecitveType, Date forcedDate) {
+		PerspectiveInterval perspectiveInterval = intervalManager
+				.getInterval(PerspectiveInterval.class);
+		if (perspectiveInterval != null && perspectiveInterval
+				.getPerspectiveType() == perspecitveType) {
+			// abort if such an interval is already open.
+			return;
+		}
+		intervalManager.closeInterval(perspectiveInterval, forcedDate);
+		intervalManager.addInterval(
+				new PerspectiveInterval(perspecitveType, forcedDate));
+	}
 
 }
